auth: fix the developer experience
Go through the normal OIDC flow since we already have the helper to
massage the auth hash into something that looks like OIDC. Move it
into a concern because that's where bad code goes to die.
