+++ app/controllers/concerns/developer_oidc.rb
@@ -1,0 +1,46 @@
+# frozen_string_literal: true
+
+module DeveloperOidc
+  extend ActiveSupport::Concern
+
+  def oidcize_dev_hash(attrs)
+    attrs.merge!(
+      {
+        **provider_info(attrs),
+        **static_info,
+        **extra_info(attrs)
+      }
+    )
+  end
+
+  private
+
+  def static_info
+    {
+      "credentials" => {
+        "token" => "dev token"
+      },
+      "uid" => "developer",
+      "info" => {
+        "name" => "Aplypo Dev",
+        "email" => "aplypro-dev-#{rand(1000)}@beta.gouv.fr"
+      }
+    }
+  end
+
+  def provider_info(attrs)
+    { "provider" => attrs["info"]["provider"] }
+  end
+
+  def extra_info(attrs)
+    uai = attrs["info"]["uai"]
+
+    {
+      extra: {
+        raw_info: {
+          FrEduRneResp: ["#{uai}$UAJ$PU$N$T3$LYC$340"]
+        }
+      }
+    }
+  end
+end
