+++ app/models/principal.rb
@@ -19,31 +19,6 @@ class Principal < ApplicationRecord
         principal.email = attrs["info"]["email"]
       end
     end
-
-    def from_developer(attrs)
-      from_oidc(oidcize_dev_hash(attrs)).tap do |principal|
-        principal.establishment = Establishment.find_or_create_by!(uai: attrs["info"]["uai"])
-      end
-    end
-
-    private
-
-    # makes the developer auth hash OIDC like
-    def oidcize_dev_hash(attrs)
-      attrs.merge(
-        {
-          "provider" => attrs["info"]["provider"],
-          "credentials" => {
-            "token" => "dev token"
-          },
-          "uid" => "developer",
-          "info" => {
-            "name" => "Aplypo Dev",
-            "email" => "aplypro-dev@beta.gouv.fr"
-          }
-        }
-      )
-    end
   end
 
   def to_s
